cmake: Install locales to correct platform paths on Linux
We still need to keep MyApp::AddTranslation for cases when the app
is built with something other than CMake, and on other platforms than Linux.

We are also adding a TREESHEETS_RELOCATABLE_INSTALLATION CMake configuration flag
that allows installing to the old relocatable location even on Linux.

Finally, we only install the necessary `*.mo` files, not the `*.po` files or `*.bat` helper scripts.