More specific loading of the localized tutorial
First try and load the file ending with the five-letter canonical form of the current locale name, e.g. `tutorial-pt_BR.cts`.
Failing that, try and load the file ending with the two-letter canonical form of the current locale name, e.g. `tutorial-pt.cts`.
Failing that, just load `tutorial.cts`, and be done with it.