Do not re-search entire document
Only refresh document in case the search mode is switched for either
"darken non-matching cell" or "highlight matches with red text color".
This way, the selection is preserved and the next match functionality
does not start from the top.

This also applies to the case sensitive search toggle.
Refreshing sufficies to evaluate the text whether it matches the search
strings and it renders accordingly.