Introduce `ThreadPool` and use it for `Display()`
Rendering the image files to bitmaps is a rather heavy task
that consumes a lot of CPU time (relatively).

Instead of rendering the images one by one,
do this in parallel with threads using the available CPU cores.

The threads are managed by `ThreadPool` that is introduced in this commit
as the base.

`ThreadPool` has been taken from [Lobster](
https://github.com/aardappel/lobster/blob/master/dev/include/ThreadPool/ThreadPool.h)