Avoid doing `LoadImage` twice for at least PNG files
The parser logic for the error case is now used for the normal parsing
of the PNG file in order to get the correct size of the PNG.

This avoids to have (very expensive) `LoadImage` being called twice as it is
later called when the buffer is converted to wxImage.

If `LoadImage` fails there, the replacement bitmap is now generated at
that step.

For JFIF, the parser logic is a little bit more complicated, so let
`LoadImage` be called in this case.