Consider TextEdit mode when copying
If the selected cell is in TextEdit mode, then only the selected part of
the cell text shall be pasted if paste operation is done after the copy
operation all within TreeSheets.

This can be reached by making the string in `sys->clipboardcopy` different
(leaving as `wxEmptyString`) from the string in the system clipboard so that
TreeSheets will not take the complete cell from the internal cell clipboard.