Put clipboard data on the system stack
The clipboard data structures are created by the first document on the
heap and then are subsequently overridden by other documents.

It can be a better design to put the clipboard data into one central place
on the stack of `System` as it is available during the lifetime of the
program.