Avoid wxDataObjectComposite for pasting
wxDataObjectComposite does not work properly on Wayland for text atoms,
so avoid it.

In addition to that, the paste operation wxDataObjects for text, image and
file name can be handled completely on the stack.

Only the drag and drop operation still requires an composite data object
on the heap, so handle it separately.