Keep track of images undergoing manipulation
Prior to this commit, the image operation got executed on each image in
the cells of the selection per cell.

The issue was that if the same image was referenced in multiple cells in
the selection, the operation was executed per cell, thus on the same image
multiple times.

In order to prevent this, keep track of all images and the cells
referring to them in the selection, to execute the image operation
only once on each image and to reset the layout of the referencing cells.