Do not use minimal rectangle when we scroll to selection (#1101)
Before this commit, the TSCanvas used the minimal rectangle when we scrolled to selection of a cell, but
when the user navigated by key arrows afterwards over a thin selection, the full rectangle was returned and
the conditions were met to scroll further. After this, with another cell selected, the minimal rectangle was
returned, but no more a condition to scroll was triggered.
This behavior caused an inconsistent user experience when e.g. the user scrolls in and out and navigates afterwards with the arrow keys.